requirements.txt
azure-common==1.1.21
azure-mgmt-resource==2.2.0
